+++ src/test/java/org/bitbuckets/arm/ForwardKinematicsTest.java
@@ -1,5 +1,6 @@
 package org.bitbuckets.arm;
 
+import org.bitbuckets.arm.kinematics.ForwardKinematics;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
