+++ src/test/java/org/bitbuckets/arm/InverseKinematicsTest.java
@@ -1,5 +1,6 @@
 package org.bitbuckets.arm;
 
+import org.bitbuckets.arm.kinematics.InverseKinematics;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
